Move StatisticsDashboard into the right directory
