+++ src/snapshots-app/client/bundles/components/submission/tabs/summary/StatisticsDashboard.jsx
